Move invocations and storage_class to interfaces
